🐛 fix(client): hoist state location to parent component
상태 정보 공유를 위해 info계층에 있던 상태를 ChatRoomTemplate 단위로 올렸습니다.
